wip: refactored dashboard settings
