+++ apps/studio/components/ui/QueryBlock/EditQueryButton.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'common'
-import { useIsInlineEditorEnabled } from 'components/interfaces/Account/Preferences/InlineEditorSettings'
+import { useIsInlineEditorEnabled } from 'components/interfaces/Account/Preferences/useDashboardSettings'
 import useNewQuery from 'components/interfaces/SQLEditor/hooks'
 import { DiffType } from 'components/interfaces/SQLEditor/SQLEditor.types'
 import { SIDEBAR_KEYS } from 'components/layouts/ProjectLayout/LayoutSidebar/LayoutSidebarProvider'
