+++ apps/studio/pages/account/me.tsx
@@ -2,8 +2,8 @@ import { AccountConnections } from 'components/interfaces/Account/Preferences/Ac
 import { AccountDeletion } from 'components/interfaces/Account/Preferences/AccountDeletion'
 import { AccountIdentities } from 'components/interfaces/Account/Preferences/AccountIdentities'
 import { AnalyticsSettings } from 'components/interfaces/Account/Preferences/AnalyticsSettings'
+import { DashboardSettings } from 'components/interfaces/Account/Preferences/DashboardSettings'
 import { HotkeySettings } from 'components/interfaces/Account/Preferences/HotkeySettings'
-import { InlineEditorSettings } from 'components/interfaces/Account/Preferences/InlineEditorSettings'
 import { ProfileInformation } from 'components/interfaces/Account/Preferences/ProfileInformation'
 import { ThemeSettings } from 'components/interfaces/Account/Preferences/ThemeSettings'
 import AccountLayout from 'components/layouts/AccountLayout/AccountLayout'
@@ -89,7 +89,7 @@ const ProfileCard = () => {
 
         <HotkeySettings />
 
-        <InlineEditorSettings />
+        <DashboardSettings />
 
         {profileShowAnalyticsAndMarketing && <AnalyticsSettings />}
